fixed several things and updated other
made it so that the functions have started to use the class string.

fixed how the constructor works to allow the string it is inputted with work with the other functions.

added a delete and nullptr to the main().

added a copy cosntructor.

added deletes and nullptrs for char containers.